requirements.txt
websocket-client>=1.5.0
PyYAML>=6.0
requests>=2.28.0
python-binance